chore(tile-debug): split logic between multiple files
