+++ src/constants/bounding-volumes.ts
@@ -1,0 +1,2 @@
+export const OBB = "Oriented Bounding Box";
+export const MBS = "Minimum Bounding Sphere";
